Update schema to newest version of the Fiscal data package spec
Budget data package has been renamed to Fiscal data package and
the license confusion has be rectified both for fiscal data
packages and for data packages so we can just reference "license".

Dimension types have also been added to the spec. Unfortunately,
the spec is inconsistent in what types it expects. The example in
the specification which the previous schema was based on names
"program" but the dimension types listed in the spec text (not in
the example) do not include "program" but contains "project"
instead (and "datetime"). We will go with what the spec text says
and not the example.

Also, the primaryKey for fields in dimensions is changed. Turns out
the schema allows defining primaryKey in two ways. Either by
having a primaryKey property with a boolean value of true or a
string value with an empty string (which in many circumstances is
interpreted as false). To support this we just go with anyOf for
the primaryKey property. Content validation is still left to the
software using the schema (we could put a enum with value [""]
but that just looks very weird and even more confusing than the
patternProperties. The spec is also not very clear on this because
the empty string is only mentioned in an example comment:

'# note being true is optional it can just be set to ""'

So by the looks of it, it's only checking for the key's existence
and not its value. Let's go with this for now.