Modify dataset title to be derived from fiscal data package metadata
We do not know what the data package title is but we want the CKAN
instance's datasets to have understandable titles. This could be
done by modifying CKAN but instead of modifying CKAN to display the
dataset title differently everywhere it's displayed we just modify
it on input and have bdp2ckan modify it. This should be handled by
a CKAN extension that parses and loads fiscal data packages (which
will replace this script but we do it like this now as a proof of
concept).